+++ lib/theme/editor.ts
@@ -21,12 +21,12 @@ export const EDITOR_CHAR_COLORS = new Uint8Array([
   ...sgr(["fg", ...LIGHTEST]),
 ]);
 
-export const EDITOR_WHITESPACE_OFF_COLORS = new Uint8Array([
+export const EDITOR_EMPTY_COLORS = new Uint8Array([
   ...EDITOR_BG,
   ...sgr(["fg", ...LOWER]),
 ]);
 
-export const EDITOR_WHITESPACE_ON_COLORS = new Uint8Array([
+export const EDITOR_WHITESPACE_COLORS = new Uint8Array([
   ...EDITOR_BG,
   ...sgr(["fg", ...DARK]),
 ]);
@@ -36,7 +36,7 @@ export const EDITOR_SELECTED_CHAR_COLORS = new Uint8Array([
   ...sgr(["fg", ...LIGHTEST]),
 ]);
 
-export const EDITOR_SELECTED_INVISIBLE_COLORS = new Uint8Array([
+export const EDITOR_SELECTED_WHITESPACE_COLORS = new Uint8Array([
   ...sgr(["bg", ...TOP]),
   ...sgr(["fg", ...DARKER]),
 ]);